Придумал как сделать определение нужных полос для создания пути, добавил глобальные параметры, продвинулся в разработке системы просчитывания клеток, полос, машин и т.д

+++ ts_server/src/main/java/com/traffic_simulator/businnes_logic/models/GraphObject.java
@@ -11,10 +11,10 @@ import java.util.*;
 @ToString
 public abstract class GraphObject {
     protected enum ElementColor {WHITE, GRAY, BLACK}
-    protected int naturalWeight;
-    protected Map<Integer, Integer> trafficWeight;
+    protected double naturalWeight;
+    protected Map<Integer, Double> trafficWeight;
     protected ElementColor mark;
-    protected int markWeight;
+    protected double markWeight;
     public GraphObject() {
         this.naturalWeight = 0;
         this.trafficWeight = new HashMap<>();
